fix the closing session bug, even if next user authentication is failed.
When switching user by calling switchUser method, current opened session
is closed without verifying the next user authentication (i.e. username
and password). Because of this reason, there could be no remains session
in case of failure of user authentication for next user.

This patch checks the next user authentication and immediately return
FALSE when failed so that current session is not terminated.

Change-Id: I794243faa7ebaf47f53bdf9b9368e14dbf8c509a
Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>